requirements.txt
aiohttp==3.5.4
aiostream==0.3.1
brotlipy==0.7.0
bs4==0.0.1
emoji==0.5.2
jmespath==0.9.4
lxml==4.3.4
python-dateutil==2.8.0
Scrapy==1.6.0
ujson==1.35
uvloop==0.12.2
wrapt==1.11.2
youtube-dl
